+++ src/app/models/automations.ts
@@ -26,10 +26,13 @@ export type AutomationType =
   | 'AUTO_ACCEPT_INVITE_REQUESTS'
   // BRIGHTNESS AUTOMATIONS
   | 'DISPLAY_BRIGHTNESS_ON_SLEEP_MODE_ENABLE'
-  | 'DISPLAY_BRIGHTNESS_ON_SLEEP_MODE_DISABLE';
+  | 'DISPLAY_BRIGHTNESS_ON_SLEEP_MODE_DISABLE'
+  // RESOLUTION AUTOMATIONS
+  | 'RENDER_RESOLUTION_ON_SLEEP_MODE_ENABLE'
+  | 'RENDER_RESOLUTION_ON_SLEEP_MODE_DISABLE';
 
 export interface AutomationConfigs {
-  version: 8;
+  version: 9;
   GPU_POWER_LIMITS: GPUPowerLimitsAutomationConfig;
   MSI_AFTERBURNER: MSIAfterburnerAutomationConfig;
   // SLEEP MODE AUTOMATIONS
@@ -52,6 +55,9 @@ export interface AutomationConfigs {
   // BRIGHTNESS AUTOMATIONS
   DISPLAY_BRIGHTNESS_ON_SLEEP_MODE_ENABLE: DisplayBrightnessOnSleepModeAutomationConfig;
   DISPLAY_BRIGHTNESS_ON_SLEEP_MODE_DISABLE: DisplayBrightnessOnSleepModeAutomationConfig;
+  // RESOLUTION AUTOMATIONS
+  RENDER_RESOLUTION_ON_SLEEP_MODE_ENABLE: RenderResolutionOnSleepModeAutomationConfig;
+  RENDER_RESOLUTION_ON_SLEEP_MODE_DISABLE: RenderResolutionOnSleepModeAutomationConfig;
 }
 
 export interface AutomationConfig {
@@ -68,6 +74,10 @@ export interface DisplayBrightnessOnSleepModeAutomationConfig extends Automation
   brightness: number;
   transition: boolean;
   transitionTime: number;
+}
+
+export interface RenderResolutionOnSleepModeAutomationConfig extends AutomationConfig {
+  resolution: number | null;
 }
 
 // GPU AUTOMATIONS
@@ -165,7 +175,7 @@ export interface AutoAcceptInviteRequestsAutomationConfig extends AutomationConf
 //
 
 export const AUTOMATION_CONFIGS_DEFAULT: AutomationConfigs = {
-  version: 8,
+  version: 9,
   // BRIGHTNESS AUTOMATIONS
   DISPLAY_BRIGHTNESS_ON_SLEEP_MODE_ENABLE: {
     enabled: false,
@@ -178,6 +188,15 @@ export const AUTOMATION_CONFIGS_DEFAULT: AutomationConfigs = {
     brightness: 100,
     transition: true,
     transitionTime: 10000,
+  },
+  // RESOLUTION AUTOMATIONS
+  RENDER_RESOLUTION_ON_SLEEP_MODE_ENABLE: {
+    enabled: false,
+    resolution: 50,
+  },
+  RENDER_RESOLUTION_ON_SLEEP_MODE_DISABLE: {
+    enabled: false,
+    resolution: null,
   },
   // GPU AUTOMATIONS
   GPU_POWER_LIMITS: {