+++ src/app/services/brightness-control/drivers/valve-index-brightness-control-driver.ts
@@ -1,5 +1,4 @@
 import { BrightnessControlDriver } from './brightness-control-driver';
-import { invoke } from '@tauri-apps/api';
 import { clamp, ensurePrecision, lerp } from '../../../utils/number-utils';
 import { OpenVRService } from '../../openvr.service';
 import { combineLatest, map, Observable } from 'rxjs';
@@ -14,7 +13,7 @@ export class ValveIndexBrightnessControlDriver extends BrightnessControlDriver {
   }
 
   async getBrightnessPercentage(): Promise<number> {
-    let analogGain = await invoke<number>('openvr_get_analog_gain');
+    let analogGain = await this.openvr.getAnalogGain();
     analogGain = ensurePrecision(analogGain, 3);
     return this.analogGainToPercentage(analogGain);
   }
@@ -23,7 +22,7 @@ export class ValveIndexBrightnessControlDriver extends BrightnessControlDriver {
     const bounds = await this.getBrightnessBounds();
     percentage = clamp(percentage, bounds[0], bounds[1]);
     const analogGain = this.percentageToAnalogGain(percentage);
-    await invoke('openvr_set_analog_gain', { analogGain });
+    this.openvr.setAnalogGain(analogGain);
   }
 
   isAvailable(): Observable<boolean> {