+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Display brightness automations for Valve Index.
+- Automations for adjusting SteamVR's render resolution.
 - Event log for displaying actions taken by Oyasumi
 - Language support for French, thanks to [neuroblack](https://github.com/neuroblack).